added multiprocess filtering and changed kill class to find terminated child processes

+++ heliplot/kill.py
@@ -27,12 +27,12 @@ class Kill(object):
 		# raise subprocess.TimeoutExpired(proc.args, output=out)
 		# os.killpg(proc.pid, signum)	# kill process group
 		# ---------------------------------------
-		print "Killing child:	%6s" % childpid
+		print "Child proc:	%6s" % childpid
 		print "Killing gchild:	%6s" % gchildpid 	
 		time.sleep(1)
 		# Send kill signal to child/grandchild (terminate/kill)
-		os.kill(gchildpid, signum)	# kill grandchild process
-		os.kill(childpid, signum)	# kill child process
+		os.kill(gchildpid, signum)	# kill grandchild process	
+		#os.kill(childpid, signum)	# kill child process
 		time.sleep(1)
 
 	def killPool(self, pid, name):
@@ -42,17 +42,21 @@ class Kill(object):
 		# os.wait()	# wait for all threads to exit
 		# pool.terminate()
 		# pool.join()
+		# pkill -TERM -P 'pid' - kill process group
+		# kill -- -$PGID - kill parent/child/gchild
 		# ---------------------------------------
-		time.sleep(1)
 		print "Pool %s is terminated" % name
 		parent = psutil.Process(pid)
 		print "Parent pid:	%6s" % pid
 		print "Killing children of pool: %6s..." % pid
-		time.sleep(1)
+		#time.sleep(1)
 		# set recursive=True to kill grandchildren
 		for child in parent.get_children(recursive=True):
-			print "\tchild: " + str(child)
-			child.kill()	# kill children of pool
+			if 'terminated' in str(child):
+				print "\tterminated child: " + str(child)
+			else:	
+				print "\tkilling child: " + str(child)
+				child.kill()	# kill children of pool
 			time.sleep(1)	
 		print "Killing pool:	%6s..." % pid
 		time.sleep(1)
